sender/pyproject.toml
[tool.poetry]
name = "gcode_robot"
version = "0.1.0"
description = ""
authors = ["Grayson Wen <wenoptics@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
